ml_pipeline_debugger/server/requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.115.0
uvicorn>=0.24.0
torch>=2.0.0
numpy>=1.24.0
openai>=1.0.0
